Fixed 3 attempt login
The program would immediately exit if a user failed to login. Changed the Auth class to have a login_attempts variable that tracks total attempts. If 3 times, closes bank and says good bye. I made it 3 times as thats what the assignment asked for. Changed Structure of banking.py also to simplify